pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-utils"
version = "0.1.0"
description = "A collection of Python utilities"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "xzsean666", email = "xz.sean.g@gmail.com"},
]
maintainers = [
    {name = "xzsean666", email = "xz.sean.g@gmail.com"},
]
keywords = ["utilities", "python", "database", "cache", "sqlite"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=1.4.0",
    "cryptography>=41.0.0",
    "PyJWT>=2.8.0",
    "pyotp>=2.9.0",
    "qrcode[pil]>=7.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12.0",
    "black>=21.0",
    "flake8>=3.9.0",
    "mypy>=0.900",
]

[project.urls]
Homepage = "https://github.com/xzsean666/py-utils"
Repository = "https://github.com/xzsean666/py-utils.git"
Issues = "https://github.com/xzsean666/py-utils/issues"
Documentation = "https://github.com/xzsean666/py-utils#readme"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
